New UI: Use uint as extension for enum;

+++ Source/GUIv2/clipmeter/ClipMeter.cpp
@@ -15,7 +15,7 @@ namespace gui
 {
 namespace
 {
-enum class ScalingType : u_int8_t
+enum class ScalingType : unsigned int
 {
     kLinear,
     kLogarithmic
