+++ Source/GUIv2/levelmeter/LevelMeterComponent.cpp
@@ -10,7 +10,7 @@ namespace gui
 {
 namespace
 {
-enum class ScalingType : u_int8_t
+enum class ScalingType : unsigned int
 {
     kLinear,
     kLogarithmic
